move resolveImages out of controllers/dynakube
